P2 dbengine: guard LZ4 size conversions
Finding MEM-dbengine-compression-003 exposed a latent contract mismatch: dbengine_decompress() accepts size_t lengths but narrows them to the int parameters required by LZ4. The sole current caller validates extents to a 502272-byte nominal uncompressed cap, so valid data cannot trigger it; a future caller or extent-limit change could pass an unrepresentable size.

Reject LZ4 source or destination sizes above INT_MAX before the casts, log the sizes, and return the existing decompression failure sentinel that the caller maps to have_read_error. Keep the always-active compression bound assertion unchanged: its largest accepted input has a 2122219149-byte bound, below INT_MAX.

Also remove the redundant size_t-to-int-to-size_t cast around dbengine_compress() in datafile_extent_build(). Valid compression bytes, fallback, disk format, CRC/accounting, ownership, lifetimes, locking, concurrency, and caller error contracts are unchanged. LZ4 reads gain one unlikely two-comparison branch; writes gain no work.

Validated with focused optimized object builds, strict format-warning compilation, an ASan/UBSan actual-source boundary probe, a full optimized/LTO daemon link, and the isolated full internal suite including DBEngine. All six required reviewers completed and approved.

(cherry picked from commit 47cd3fb16786e3b89c1fb765dfb6b3c4b015f7b6)

+++ src/database/engine/rrdengine.c
@@ -985,9 +985,9 @@ datafile_extent_build(struct rrdengine_instance *ctx, struct page_descr_with_dat
 
     // compress the payload
     size_t compressed_size =
-        (int)dbengine_compress(xt_io_descr->buf + payload_offset,
-                               uncompressed_payload_length,
-                               compression_algorithm);
+        dbengine_compress(xt_io_descr->buf + payload_offset,
+                          uncompressed_payload_length,
+                          compression_algorithm);
 
     internal_fatal(compressed_size > max_compressed_size, "DBENGINE: compression returned more data than the max allowed");
     internal_fatal(compressed_size > uncompressed_payload_length, "DBENGINE: compression returned more data than the uncompressed extent");
